# Avoid to log if the log is not set