`loggerUtils.onResponseCallback` logs only if there's a logger

Bench:

*this branch:*
```
[1] Running 5s test @ http://localhost:3000/
[1] 100 connections with 10 pipelining factor
[1] 
[0] server listening on 3000
[1] Stat         Avg     Stdev  Max     
[1] Latency (ms) 8.3     27.44  511.66  
[1] Req/Sec      11710   2804.9 13896   
[1] Bytes/Sec    1.75 MB 411 kB 2.07 MB 
[1] 
[1] 59k requests in 5s, 8.72 MB read
[1] 112 errors (0 timeouts)
```

*master:*
```
[1] Running 5s test @ http://localhost:3000/
[1] 100 connections with 10 pipelining factor
[1] 
[0] server listening on 3000
[1] Stat         Avg     Stdev   Max     
[1] Latency (ms) 8.78    29.27   340.12  
[1] Req/Sec      11079.6 3750.94 14926   
[1] Bytes/Sec    1.64 MB 535 kB  2.22 MB 
[1] 
[1] 55k requests in 5s, 8.25 MB read
[1] 130 errors (0 timeouts)
```

Can someone else confirm this improvement?

#### Checklist

- [x] run `npm run test` and `npm run benchmark`
- [x] tests and/or benchmarks are included
- [ ] documentation is changed or added
- [x] commit message and code follows [Code of conduct](https://github.com/fastify/fastify/blob/master/CODE_OF_CONDUCT.md)