@allevo 2018-02-13
@delvedor yeah, `_startTime` should be considered as a private variable.