@allevo 2018-02-13
New bench:
*master*
```
754k requests in 50s, 112 MB read
````
*this branch*
```
787k requests in 50s, 117 MB read
```

I don't know if this bench are accurate or not. Is there someone that is able to confirm it?