@allevo 2018-02-13
Last commit remove two listener: if the logger and the current context has no `onResponse` hooks, those lines can be skipped

https://github.com/fastify/fastify/blob/8aa6b1dfb8b7f0298c72dcf2b8aafcf003a87dfb/fastify.js#L203-L204