@allevo 2018-02-26
Tomorrow I'll update this PR

Il lun 26 feb 2018, 12:47 Matteo Collina <notifications@github.com> ha
scritto:

> @allevo <https://github.com/allevo> any update on this?
>
> —
> You are receiving this because you were mentioned.
>
>
> Reply to this email directly, view it on GitHub
> <https://github.com/fastify/fastify/pull/776#issuecomment-368475481>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/ABAVrVXTmxeOnec9TfYBlOIUkN5eOaP-ks5tYpnlgaJpZM4SBjvX>
> .
>