@mcollina 2018-02-28
A quick explanation on the `Date.now()` problem:
 
![schermata 2018-02-28 alle 14 31 58](https://user-images.githubusercontent.com/52195/36790136-28770066-1c94-11e8-8a3d-89b2d9a90dda.png)