@allevo 2018-02-28
@mcollina Removing the call to `now`, we should remove that 5.8% on the top. Am I right?